requirements.txt
numpy
matplotlib
seaborn
torch
tqdm
nvitop
torchvision
pillow
diffusers
transformers 
accelerate 
scipy 
safetensors
bitsandbytes
# lpips
scikit-image
imagehash
deoldify
fastprogress
albumentations